復刻版カードはオリジナルをimportするだけに
