+++ src/database/effects/cards/1-0-031.ts
@@ -1,8 +1,8 @@
-import type { Stack } from '@/package/core/class/stack';
 import { EffectTemplate, System } from '..';
+import type { StackWithCard } from '../classes/types';
 
 export const effects = {
-  onOverclockSelf: async (stack: Stack) => {
+  onOverclockSelf: async (stack: StackWithCard) => {
     await System.show(stack, 'リバイブ', '捨札から1枚選んで回収');
     await EffectTemplate.revive(stack, 1);
   },
