+++ src/database/effects/cards/PR-189.ts
@@ -1,0 +1,2 @@
+import { effects } from './1-3-120';
+export { effects };
