+++ src/database/effects/cards/PR-194.ts
@@ -1,0 +1,2 @@
+import { effects } from './2-0-024';
+export { effects };
